Modify ObservePage.html and its function in afl_ contoller.py, afl_service.py that user can see the uploaded file in observePage

+++ app/models/fuzzConfig.py
@@ -1,13 +1,15 @@
 from typing import Dict
+import os
 
 # 定義 FuzzConfig 類別
 class FuzzConfig:
     def __init__(self, configParams: Dict[str, str], targetFolder: str):
         self.configParams = configParams
         self.targetFolder = targetFolder
-        self.configPersistence()
-
-    def configPersistence(self) -> None:
-        with open(f"{self.targetFolder}/config.txt", "w") as file:
-            for key, value in self.configParams.items():
-                file.write(f"{key}: {value}\n")
+        self.setConfigPersistence()
+        
+    def setConfigPersistence(self) -> None:
+        if not os.path.exists(f"{self.targetFolder}/config.txt"):
+            with open(f"{self.targetFolder}/config.txt", "w") as file:
+                for key, value in self.configParams.items():
+                    file.write(f"{key}: {value}\n")
